JVM_IR: Move suspend function views creation to lowering
Now AddContinuationLowering is responsible for both adding continuation
classes to suspend functions and and adding continuation parameters to
them.
Because we cannot create view if inline suspend function is defined in
another class, we generate stubs without bodies when we encounter calls
to them. And then, when we lower the file containing the function we
add body to it. This way we have no unlowered views after the lowering.
Because after the lowering there should be no suspend function, which
are not views, we remove VIEW origins.
Because transformations of suspend functions can copy them into another
object, use attribute as a key inside function to view map.